pyproject.toml
[project]
name = "imti"
version = "0.1.0"
description = "India Market Timing Index — AI-powered market timing signal system"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "IMTI Team" }]
dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "requests>=2.32",
    "beautifulsoup4>=4.12",
    "yfinance>=0.2.40",
    "xgboost>=2.1",
    "scikit-learn>=1.5",
    "shap>=0.46",
    "matplotlib>=3.9",
    "plotly>=5.22",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "sendgrid>=6.11",
    "openai>=1.35",
    "lxml>=5.2",
    "httpx>=0.27",
    "tenacity>=8.5",
    "pydantic>=2.8",
    "ta>=0.11",
    "pynacl>=1.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.11",
    "types-requests>=2.32",
    "types-pyyaml>=6.0",
    "types-beautifulsoup4>=4.12",
]

[project.scripts]
imti = "imti.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/imti"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["imti"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
